requirements.txt
Django>=4.2.0
Pillow>=10.0.0
qrcode[pil]>=7.0.0
python-barcode>=0.15.0
reportlab>=4.0.0
requests>=2.31.0
sqlparse>=0.4.0
python-whois>=0.8.0
PyMuPDF>=1.23.0
cairosvg>=2.7.0
opencv-python>=4.8.0
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=2.0.0